feat(blog): refactor comment composer and related components, remove unused blog reads info
