+++ src/app/(site)/blog/[[...slug]]/comment-card.tsx
@@ -2,7 +2,7 @@
 
 import type { CommentWithMeta } from "./actions";
 import { Avatar } from "./avatar";
-import { CommentComposer } from "./comment-composer";
+import { CommentComposer } from "./comment-composer-area";
 import {
   ArrowBendDownRightIcon,
   HeartIcon,
